# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI MVP API
  version: 0.2.0
  description: BuildGraph 3주 MVP 공통 API 계약. 상세 기준은 docs/API_CONTRACT.md를 따른다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/users:
    post:
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: 생성된 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/auth/login:
    post:
      summary: 이메일 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT와 사용자 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      summary: access token 재발급
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: 재발급된 token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
  /api/auth/logout:
    post:
      security:
        - bearerAuth: []
      summary: 로그아웃
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: 로그아웃 완료
  /api/auth/me:
    get:
      summary: 현재 사용자 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 현재 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /api/ai/as-chat:
    get:
      summary: AS AI 챗봇 세션과 메시지 조회
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: asTicketId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: AS AI 챗봇 세션과 대화 이력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsChatResponse"
    post:
      summary: AS AI 챗봇 메시지 전송
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-BuildGraph-AI-Profile
          required: false
          description: 내부 benchmark용 profile override. 사용자 화면은 보내지 않는다.
          schema:
            type: string
            enum: [AS_CHAT_FAST, AS_CHAT_NANO_FAST, AS_CHAT_BALANCED, AS_CHAT_HIGH_QUALITY]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsChatRequest"
      responses:
        "200":
          description: AS AI 챗봇 LLM 구조화 응답
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsChatResponse"
        "428":
          description: OPENAI_API_KEY 등 서버 선행 설정 누락
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: LLM JSON 계약 위반 또는 외부 LLM 응답 처리 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/ai/as-chat/stream:
    post:
      summary: AS AI 챗봇 메시지 전송 진행 상태 SSE
      description: STARTED, RAG_READY, TOOLS_READY, LLM_RUNNING, DONE, ERROR 이벤트를 text/event-stream으로 반환한다. DONE data는 AsChatResponse JSON이다.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: X-BuildGraph-AI-Profile
          required: false
          description: 내부 benchmark용 profile override. 사용자 화면은 보내지 않는다.
          schema:
            type: string
            enum: [AS_CHAT_FAST, AS_CHAT_NANO_FAST, AS_CHAT_BALANCED, AS_CHAT_HIGH_QUALITY]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsChatRequest"
      responses:
        "200":
          description: AS AI 챗봇 진행 상태와 최종 구조화 응답 SSE
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: USER 인증 필요
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT_STATE, DUPLICATE_RESOURCE, FILE_VALIDATION_ERROR, PRECONDITION_REQUIRED, UPSTREAM_ERROR, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Role:
      type: string
      enum: [USER, ADMIN]
...
    ToolStatus:
      type: string
      enum: [PASS, WARN, FAIL]
    Confidence:
      type: string
      enum: [LOW, MEDIUM, HIGH]
...
    AsTicketStatus:
      type: string
      enum: [OPEN, ASSIGNED, IN_PROGRESS, RESOLVED, CLOSED, CANCELLED]
...
    UserDto:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    SignupRequest:
      type: object
      required: [email, password, name, termsAccepted, marketingAccepted]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        termsAccepted:
          type: boolean
        marketingAccepted:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserDto"
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
    RagEvidenceDto:
      type: object
      required: [id, sourceId, summary]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
        summary:
          type: string
        score:
          type: number
        metadata:
          type: object
          additionalProperties: true
...
    AsChatRequest:
      type: object
      required: [asTicketId, message]
      properties:
        asTicketId:
          type: string
          format: uuid
        message:
          type: string
    AsChatResponse:
      type: object
      required: [asTicketId, ticket, model, messages, evidence, toolResults]
      properties:
        sessionId:
          type: string
          format: uuid
          nullable: true
        asTicketId:
          type: string
          format: uuid
        ticket:
          $ref: "#/components/schemas/AsChatTicketDto"
        model:
          type: string
        agentSessionId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AsChatMessageDto"
        assistantMessage:
          type: string
        causeCandidates:
          type: array
          items:
            $ref: "#/components/schemas/AsChatCauseCandidateDto"
        nextActions:
          type: array
          items:
            $ref: "#/components/schemas/AsChatNextActionDto"
        escalation:
          $ref: "#/components/schemas/AsChatEscalationDto"
        ticketDraft:
          $ref: "#/components/schemas/AsChatTicketDraftDto"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/AsChatEvidenceDto"
        toolResults:
          type: array
          items:
            $ref: "#/components/schemas/AsChatToolResultDto"
    AsChatTicketDto:
      type: object
      required: [id, status, symptom]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AsTicketStatus"
        symptom:
          type: string
        logSummary:
          type: string
        causeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        upgradeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        createdAt:
          type: string
          format: date-time
    AsChatMessageDto:
      type: object
      required: [id, role, content]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [USER, ASSISTANT]
        content:
          type: string
        structuredPayload:
          type: object
          additionalProperties: true
        agentSessionId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    AsChatCauseCandidateDto:
      type: object
      required: [label, confidence, reason, evidenceIds, toolInvocationIds]
      properties:
        label:
          type: string
        confidence:
          $ref: "#/components/schemas/Confidence"
        reason:
          type: string
        evidenceIds:
          type: array
          items:
            type: string
            format: uuid
        toolInvocationIds:
          type: array
          items:
            type: string
            format: uuid
    AsChatNextActionDto:
      type: object
      required: [label, priority, instruction, evidenceIds, toolInvocationIds]
      properties:
        label:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        instruction:
          type: string
        evidenceIds:
          type: array
          items:
            type: string
            format: uuid
        toolInvocationIds:
          type: array
          items:
            type: string
            format: uuid
    AsChatEscalationDto:
      type: object
      required: [required, reason]
      properties:
        required:
          type: boolean
        reason:
          type: string
    AsChatTicketDraftDto:
      type: object
      required: [symptomSummary, recommendedLogRequest]
      properties:
        symptomSummary:
          type: string
        recommendedLogRequest:
          type: string
    AsChatEvidenceDto:
      allOf:
        - $ref: "#/components/schemas/RagEvidenceDto"
        - type: object
          properties:
            chunkText:
              type: string
    AsChatToolResultDto:
      type: object
      required: [id, toolName, status, confidence, summary]
      properties:
        id:
          type: string
          format: uuid
        toolName:
          type: string
          enum: [compatibility, power, size, performance, price]
        status:
          $ref: "#/components/schemas/ToolStatus"
        confidence:
          $ref: "#/components/schemas/Confidence"
        summary:
          type: string
        resultPayload:
          type: object
          additionalProperties: true
...